+++ libs/filabridge/include/filament/MaterialChunkType.h
@@ -50,6 +50,7 @@ enum UTILS_PUBLIC ChunkType : uint64_t {
 
     MaterialName = charTo64bitNum("MAT_NAME"),
     MaterialVersion = charTo64bitNum("MAT_VERS"),
+    MaterialFeatureLevel = charTo64bitNum("MAT_FEAT"),
     MaterialShading = charTo64bitNum("MAT_SHAD"),
     MaterialBlendingMode = charTo64bitNum("MAT_BLEN"),
     MaterialTransparencyMode = charTo64bitNum("MAT_TRMD"),